# docs/openapi.yaml
openapi: "3.0.3"
info:
  title: Viper Chain API
  version: "0.2.0"
  description: >
    API for interacting with Viper PQ Chain — a post-quantum digital trust
    L1. The public chain is `viper-testnet-1` (`chain_id_hex` is the hex of
    the ASCII chain id, e.g. `76697065722d746573746e65742d31`); it is
    created after the public release, so there is no live network to
    query at publication time. Routes marked "specified, not implemented"
    are documented for SDK alignment but are not served by `pqcd`. Use these
    endpoints to query chain state, submit transactions, look up
    attestations / proofs / governance receipts, manage verifiable
    credentials, and anchor proofs.
    Spec version aligns with the Viper SDK 0.2.0 release (TS + Python),
    which predates `viper-testnet-1` and is re-aligned before republishing.
servers:
  - url: https://rpc.pqchain.agwswebconsulting.it
    description: Planned public read API of viper-testnet-1 (DNS published at genesis; not live at publication time)
  - url: http://localhost:26657
    description: Local devnet node (default API port)
...
paths:
...
  /v1/proofs/{anchor_id}:
    get:
      tags: [Proofs]
      summary: Get proof anchor by anchor id
      description: >
        Returns a finalized proof anchor record by its 32-byte anchor
        id (64 hex characters). Implemented by the node runtime at
        `crates/pqcd/src/devnet.rs::handle_get_proof_anchor` and by the
        api-only handler at `crates/pqcd/src/api.rs::handle_proof_anchor`.
      operationId: getProofAnchor
      parameters:
        - name: anchor_id
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9a-f]{64}$"
          description: Anchor id (64 hex characters)
      responses:
        "200":
          description: Proof anchor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProofAnchorResponse"
        "400":
          description: Invalid anchor id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Proof anchor not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/proofs/anchor:
    post:
      tags: [Proofs]
      summary: Anchor a proof
      description: >
        Validates proof anchoring parameters and returns a structured response
        describing the transaction that would be built. This is a convenience
        endpoint for building proof anchor transactions.
      operationId: anchorProof
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProofAnchorRequest"
            example:
              owner_address: "abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789"
              claim_type: "ownership"
              document_hash: "aabbccdd00112233aabbccdd00112233aabbccdd00112233aabbccdd00112233"
              proof_hash: "11223344556677881122334455667788112233445566778811223344556677aa"
      responses:
        "200":
          description: Proof anchoring accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProofAnchorApiResponse"
              example:
                status: "accepted"
                proof_id: "f1e2d3c4b5a6..."
                msg_type: "proof_anchor"
                verify_url: "/api/proofs/f1e2d3c4b5a6..."
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/proofs/{id}:
    get:
      tags: [Proofs]
      summary: Get proof by id
      description: >
        Returns a proof anchor record with business-friendly field names.
        Wraps the on-chain proof anchor data.
      operationId: getProof
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9a-f]{64}$"
          description: Proof ID (64 hex characters)
      responses:
        "200":
          description: Proof details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProofResponse"
        "400":
          description: Invalid proof id format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Proof not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Governance ───────────────────────────────────────────────────────────────
...
  /api/credentials/issue:
    post:
      tags: [Credentials]
      summary: Issue a credential
      description: >
        Validates credential issuance parameters and returns a structured response
        describing the transaction that would be built. This is a convenience
        endpoint for building credential issuance transactions.
      operationId: issueCredential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CredentialIssueRequest"
            example:
              issuer_address: "abcdef0123456789abcdef0123456789abcdef0123456789abcdef0123456789"
              subject: "1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef"
              credential_type: "diploma"
              content_hash: "aabbccdd00112233aabbccdd00112233aabbccdd00112233aabbccdd00112233"
              schema_id: "edu-diploma-v1"
      responses:
        "200":
          description: Credential issuance accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialIssueResponse"
              example:
                status: "accepted"
                credential_id: "e3b0c44298fc1c149afbf4c8996fb924..."
                msg_type: "attestation_create"
                verify_url: "/api/credentials/e3b0c44298fc1c149afbf4c8996fb924..."
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/credentials/{id}:
    get:
      tags: [Credentials]
      summary: Get credential by id
      description: >
        Returns a verifiable credential record with business-friendly field names.
        Wraps the on-chain attestation data.
      operationId: getCredential
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            pattern: "^[0-9a-f]{64}$"
          description: Credential ID (64 hex characters)
      responses:
        "200":
          description: Credential details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CredentialResponse"
        "400":
          description: Invalid credential id format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Credential not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Transactions ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error description
          required: [code, message]

    # ── Chain ──────────────────────────────────────────────────────────────────
...
    ProofAnchorData:
      type: object
      properties:
        anchor_id:
          type: string
        claimer:
          type: string
        claim_type:
          type: integer
        claim_type_name:
          type: string
          nullable: true
        asset_id_hash:
          type: string
        proof_hash:
          type: string
        schema_id:
          type: string
          nullable: true
        anchor_height:
          type: integer
          format: int64
    ProofAnchorResponse:
      type: object
      description: Wrapper returned by GET /v1/proofs/{anchor_id}.
      properties:
        data:
          $ref: "#/components/schemas/ProofAnchorData"
    ProofAnchorRequest:
      type: object
      required: [owner_address, claim_type, document_hash, proof_hash]
      properties:
        owner_address:
          type: string
          pattern: "^[0-9a-f]{64}$"
          description: Owner account address (64 hex characters)
        claim_type:
          type: string
          description: Type of claim (e.g. "ownership", "custody", "asset_metadata")
        document_hash:
          type: string
          pattern: "^[0-9a-f]{64}$"
          description: Hash of the document being proven (64 hex characters)
        proof_hash:
          type: string
          pattern: "^[0-9a-f]{64}$"
          description: Hash of the proof material (64 hex characters)
    ProofAnchorApiResponse:
      type: object
      properties:
        status:
          type: string
          enum: [accepted]
        proof_id:
          type: string
          description: Computed proof identifier
        msg_type:
          type: string
          enum: [proof_anchor]
        verify_url:
          type: string
          description: URL to verify this proof
    ProofResponse:
      type: object
      properties:
        proof_id:
          type: string
        owner:
          type: string
        claim_type:
          type: string
        document_hash:
          type: string
        proof_hash:
          type: string
        anchored_at_block:
          type: integer
          format: int64
        status:
          type: string
          enum: [anchored]
        verify_url:
          type: string

    # ── Governance ─────────────────────────────────────────────────────────────
...
    CredentialIssueRequest:
      type: object
      required: [issuer_address, subject, credential_type, content_hash, schema_id]
      properties:
        issuer_address:
          type: string
          pattern: "^[0-9a-f]{64}$"
          description: Issuer account address (64 hex characters)
        subject:
          type: string
          pattern: "^[0-9a-f]{64}$"
          description: Subject identifier (64 hex characters)
        credential_type:
          type: string
          description: Human-readable credential type (e.g. "diploma", "license")
        content_hash:
          type: string
          pattern: "^[0-9a-f]{64}$"
          description: Hash of the credential content (64 hex characters)
        schema_id:
          type: string
          description: Schema identifier for the credential format
    CredentialIssueResponse:
      type: object
      properties:
        status:
          type: string
          enum: [accepted]
        credential_id:
          type: string
          description: Computed credential identifier
        msg_type:
          type: string
          enum: [attestation_create]
        verify_url:
          type: string
          description: URL to verify this credential
    CredentialResponse:
      type: object
      properties:
        credential_id:
          type: string
        issuer:
          type: string
        subject:
          type: string
        credential_type:
          type: string
        issued_at_block:
          type: integer
          format: int64
        status:
          type: string
          enum: [active, revoked]
        content_hash:
          type: string
        verify_url:
          type: string

    # ── Transactions ───────────────────────────────────────────────────────────
...